pyproject.toml
[tool.poetry]
name = "linebot-openai"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "linebot_openai"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.3.2"
line-bot-sdk = "^3.1.0"
pymongo = "^4.4.1"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
